+++ application/models/Parts.php
@@ -6,6 +6,7 @@
  * Date: 2/8/2017
  * Time: 1:01 PM
  */
+//Model for robot's parts
 class Parts extends CI_Model {
 // The data comes from http://www.quotery.com/top-100-funny-quotes-of-all-time/?PageSpeed=noscript
 
@@ -69,7 +70,7 @@ class Parts extends CI_Model {
         return md5(uniqid());
     }
 
-    // retrieve a single quote
+    // retrieve a single part
     public function get($which)
     {
         // iterate over the data until we find the one we want
@@ -79,7 +80,7 @@ class Parts extends CI_Model {
         return null;
     }
 
-    // retrieve all of the quotes
+    // retrieve all of the parts
     public function all()
     {
         return $this->data;