+++ ChangeLog.md
@@ -14,3 +14,5 @@ change log (newest last)
 * finished robot controller and model by Lei
 
 * finished TransacPartsRecords and PurchasePartsRecords model by Jing
+
+* finished parts controller and view by Jing